api/python/requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0 